Chore: Upgrade near-workspaces to version 0.9
