--- aurora-rust-sdk/Cargo.toml
@@ -1,6 +1,0 @@
-[workspace]
-members = [
-  "aurora-sdk",
-  "aurora-sdk-integration-tests",
-]
-exclude = ["target"]
