+++ aurora-rust-sdk/aurora-sdk-integration-tests/Cargo.toml
@@ -6,17 +6,16 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
-aurora-engine = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "cd4e046699a72cab0adf02bb5d7c47783732cf1b", default-features = false, features = ["std", "borsh-compat"] }
-aurora-engine-precompiles = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "cd4e046699a72cab0adf02bb5d7c47783732cf1b", default-features = false, features = ["std", "borsh-compat"] }
-aurora-engine-sdk = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "cd4e046699a72cab0adf02bb5d7c47783732cf1b", default-features = false, features = ["std", "borsh-compat"] }
-aurora-engine-transactions = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "cd4e046699a72cab0adf02bb5d7c47783732cf1b" }
-aurora-engine-types = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "cd4e046699a72cab0adf02bb5d7c47783732cf1b", default-features = false, features = ["std", "borsh-compat"] }
+aurora-engine = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "9a0058c8fd3b60fb68758e9edd6a42a7f43a0165", default-features = false, features = ["std"] }
+aurora-engine-precompiles = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "9a0058c8fd3b60fb68758e9edd6a42a7f43a0165", default-features = false, features = ["std"] }
+aurora-engine-sdk = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "9a0058c8fd3b60fb68758e9edd6a42a7f43a0165", default-features = false, features = ["std"] }
+aurora-engine-transactions = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "9a0058c8fd3b60fb68758e9edd6a42a7f43a0165" }
+aurora-engine-types = { git = "https://github.com/aurora-is-near/aurora-engine.git", rev = "9a0058c8fd3b60fb68758e9edd6a42a7f43a0165", default-features = false, features = ["std"] }
 base64 = "0.21"
 ethabi = "18"
 hex = "0.4"
-near-sdk = "4.1.0"
 serde = "1"
 serde_json = "1"
 tokio = { version = "1", features = ["full"] }
 toml = "0.5"
-workspaces = "0.7"
+workspaces = { version = "0.9", package = "near-workspaces" }
