+++ aurora-rust-sdk/rust-toolchain
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "1.69.0"
+channel = "1.74.0"
 components = ["clippy", "rustfmt"]
