+++ examples/social-db-from-aurora/integration-tests/rust-toolchain
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "1.66.1"
+channel = "1.74.0"
 components = ["clippy", "rustfmt"]
