+++ examples/tic-tac-toe/rust-toolchain
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "1.68.2"
+channel = "1.74.0"
 components = ["clippy", "rustfmt"]
