--- examples/uniswap-from-near/Cargo.toml
@@ -1,14 +1,0 @@
-[workspace]
-members = [
-  "contract",
-  "integration-tests",
-]
-exclude = ["target"]
-
-[profile.release]
-codegen-units = 1
-opt-level = "z"
-lto = true
-debug = false
-panic = "abort"
-overflow-checks = true
